Added 2 functions to the DiscoveredSnapsDBHelper. One for inserting and one for checking if a snap already exists in the database.
I put calls to these functions in the QueryPhotos class where the TODOs said the functionality was needed.
Added an imageURL to the DiscoveredContract as a way to tell if a snap already exists.
Added a DateHelper class so the timestamp format can be consistent across the project and to make it easy to get the time.
I have not tested it to see if the database works, but the changes don't break the app.